requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
coinbase==2.1.0
Django==3.2.12
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycryptodome==3.14.1
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
zipp==3.7.0
